refactor: application의 period 타입 int에서 long으로 변경

+++ src/main/java/com/pi/domain/application/application/service/ApplicationService.java
@@ -93,7 +93,7 @@ public class ApplicationService {
         fileKeys.forEach(awsS3Service::deleteFile);
     }
 
-    private Application update(Application application, String content, Long salary, Integer period) {
+    private Application update(Application application, String content, Long salary, Long period) {
         application.modify(content, salary, period);
         return application;
     }
